+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [#1855](https://github.com/openshift/cluster-monitoring-operator/pull/1855) Add nodeExporter.collectors.cpufreq settings.
 - [#1882](https://github.com/openshift/cluster-monitoring-operator/issues/1882) Allow configuring secrets in alertmanager component (platform)
 - [#1876](https://github.com/openshift/cluster-monitoring-operator/pull/1876) Add nodeExporter.collectors.tcpstat settings.
+- [#1883](https://github.com/openshift/cluster-monitoring-operator/pull/1883) Add nodeExporter.collectors.netdev settings.
 
 
 ## 4.12